+++ src/main/java/com/everypet/member/service/MemberService.java
@@ -1,10 +1,10 @@
 package com.everypet.member.service;
 
-import com.everypet.member.data.domain.Address;
-import com.everypet.member.data.dto.MemberDTO;
+import com.everypet.member.data.dto.SignupRequestDTO;
+import com.everypet.member.data.vo.Address;
 import org.springframework.security.core.userdetails.UserDetailsService;
 
 public interface MemberService extends UserDetailsService {
-    void register(MemberDTO memberDTO);
+    void register(SignupRequestDTO signupRequestDTO);
     void addressRegister(Address address);
 }
